+++ src/components/base/b-router/test/unit/transition/parameters.ts
@@ -183,4 +183,131 @@ test.describe('<b-router> passing transition parameters', () => {
 			});
 		});
 	});
+
+	test.describe('passing parameters to a route with the same name as the current one', () => {
+		test(
+			'if the route name is set to null, then the new parameters should be merged with the old ones',
+
+			async ({page}) => {
+				const root = await createInitRouter('history')(page);
+
+				await test.expect(root.evaluate(async (ctx) => {
+					const
+						router = ctx.router!,
+						transitions: Dictionary = {};
+
+					await router.push('main', {query: {foo: 1}});
+					transitions.path1 = getPath();
+
+					await router.push(null, {query: {baz: 1}});
+					transitions.path2 = getPath();
+
+					await router.push(null, {query: {bar: 2}});
+					transitions.path3 = getPath();
+
+					await router.push(null, {query: {bar: null}});
+					transitions.path4 = getPath();
+
+					return transitions;
+
+					function getPath() {
+						return location.pathname + location.search;
+					}
+
+				})).resolves.toEqual({
+					path1: '/?foo=1',
+					path2: '/?baz=1&foo=1',
+					path3: '/?bar=2&baz=1&foo=1',
+					path4: '/?baz=1&foo=1'
+				});
+			}
+		);
+
+		test(
+			[
+				'if the route name is explicitly set as a string, ',
+				'then the new parameters should completely replace the old ones'
+			].join(''),
+
+			async ({page}) => {
+				const root = await createInitRouter('history')(page);
+
+				await test.expect(root.evaluate(async (ctx) => {
+					const
+						router = ctx.router!,
+						transitions: Dictionary = {};
+
+					await router.push('main', {query: {foo: 1}});
+					transitions.path1 = getPath();
+
+					await router.push('main', {query: {baz: 1}});
+					transitions.path2 = getPath();
+
+					await router.push('second', {query: {bar: 1}});
+					transitions.path3 = getPath();
+
+					await router.push('second', {query: {}});
+					transitions.path4 = getPath();
+
+					return transitions;
+
+					function getPath() {
+						return location.pathname + location.search;
+					}
+
+				})).resolves.toEqual({
+					path1: '/?foo=1',
+					path2: '/?baz=1',
+					path3: '/second-page?bar=1',
+					path4: '/second-page'
+				});
+			}
+		);
+
+		test(
+			'parameters should never be merged if we are navigating through the history',
+
+			async ({page}) => {
+				const root = await createInitRouter('history')(page);
+
+				await test.expect(root.evaluate(async (ctx) => {
+					const
+						router = ctx.router!,
+						transitions: Dictionary = {};
+
+					await router.push('main', {query: {foo: 1}});
+					transitions.path1 = getPath();
+
+					await router.push(null, {query: {baz: 1}});
+					transitions.path2 = getPath();
+
+					await router.push('main', {query: {baz: 2}});
+					transitions.path3 = getPath();
+
+					await router.back();
+					transitions.path4 = getPath();
+
+					await router.back();
+					transitions.path5 = getPath();
+
+					await router.forward();
+					transitions.path6 = getPath();
+
+					return transitions;
+
+					function getPath() {
+						return location.pathname + location.search;
+					}
+
+				})).resolves.toEqual({
+					path1: '/?foo=1',
+					path2: '/?baz=1&foo=1',
+					path3: '/?baz=2',
+					path4: '/?baz=1&foo=1',
+					path5: '/?foo=1',
+					path6: '/?baz=1&foo=1'
+				});
+			}
+		);
+	});
 });