+++ src/core/component/decorators/interface/prop.ts
@@ -122,7 +122,7 @@ export interface DecoratorProp<
 	/**
 	 * If set to true, the prop always uses its own default value when needed.
 	 * This option is actually used when the `defaultProps` property is set to false for the described component
-	 * (via the `@component` decorator) and we want to override this behavior for a particular prop.
+	 * (via the `@component` decorator), and we want to override this behavior for a particular prop.
 	 *
 	 * @default `false`
 	 */
@@ -149,7 +149,8 @@ export interface DecoratorProp<
 	 *
 	 *     // Also, see core/object/watch
 	 *     {
-	 *       // If set to false, then a handler that is invoked on the watcher event does not take any arguments from the event
+	 *       // If set to false,
+	 *       // then a handler that is invoked on the watcher event does not take any arguments from the event
 	 *       provideArgs: false,
 	 *
 	 *       // How the event handler should be called: