+++ src/components/base/b-router/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v4.?.? (2023-??-??)
+
+#### :boom: Breaking Change
+
+* Merging of parameters when navigating to a route with the same name as the current one has been removed
+
 ## v4.0.0-beta.8 (2023-07-19)
 
 #### :rocket: New Feature